final project tutorial

+++ modules/Tutorial/Category/Http/Controllers/CategoryController.php
@@ -4,8 +4,9 @@ namespace Tutorial\Category\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use Tutorial\Category\Http\Requests\CategoryRequest;
+use Tutorial\Category\Models\Category;
 use Tutorial\Category\Repositories\CategoryRepo;
-use Tutorial\Category\Responses\AjaxResponses;
+use Tutorial\Common\Responses\AjaxResponses;
 
 class CategoryController extends Controller
 {
@@ -18,18 +19,21 @@ class CategoryController extends Controller
 
     public function index()
     {
+        $this->authorize('manage', Category::class);
         $categories = $this->repo->all();
         return view('Category::index', compact('categories'));
     }
 
     public function store(CategoryRequest $request)
     {
+        $this->authorize('manage', Category::class);
         $this->repo->store($request);
         return back();
     }
 
     public function edit($categoryId)
     {
+        $this->authorize('manage', Category::class);
         $categories = $this->repo->allExpectById($categoryId);
         $category = $this->repo->findById($categoryId);
         return view('Category::edit', compact('categories', 'category'));
@@ -37,11 +41,14 @@ class CategoryController extends Controller
 
     public function update($categoryId, CategoryRequest $request)
     {
-        $this->repo->update($categoryId,$request);
+        $this->authorize('manage', Category::class);
+        $this->repo->update($categoryId, $request);
         return redirect(route('categories.index'));
     }
 
-    public function destroy($categoryId){
+    public function destroy($categoryId)
+    {
+        $this->authorize('manage', Category::class);
         $this->repo->delete($categoryId);
         return AjaxResponses::SuccessResponse();
     }
